adjusted minespace renderRadioButtons to typescript with boolean option values
